snap: move confinement types into types.
